Place ifdefs around debug Serial.prints
